+++ src/main/java/org/cojen/tupl/BoundedCursor.java
@@ -433,6 +433,11 @@ final class BoundedCursor implements Cursor {
     }
 
     @Override
+    public LockResult lock() throws IOException {
+        return mSource.lock();
+    }
+
+    @Override
     public LockResult load() throws IOException {
         return mSource.load();
     }