+++ src/main/java/org/cojen/tupl/Cursor.java
@@ -216,7 +216,9 @@ public interface Cursor {
      * @throws NullPointerException if limit key is null
      * @throws IllegalStateException if position is undefined at invocation time
      */
-    public LockResult nextLe(byte[] limitKey) throws IOException;
+    public default LockResult nextLe(byte[] limitKey) throws IOException {
+        return ViewUtils.nextCmp(this, limitKey, 1);
+    }
 
     /**
      * Moves to the Cursor to the next available entry, but only when less than
@@ -230,7 +232,9 @@ public interface Cursor {
      * @throws NullPointerException if limit key is null
      * @throws IllegalStateException if position is undefined at invocation time
      */
-    public LockResult nextLt(byte[] limitKey) throws IOException;
+    public default LockResult nextLt(byte[] limitKey) throws IOException {
+        return ViewUtils.nextCmp(this, limitKey, 0);
+    }
 
     /**
      * Moves to the Cursor to the previous available entry. Cursor key and
@@ -258,7 +262,9 @@ public interface Cursor {
      * @throws NullPointerException if limit key is null
      * @throws IllegalStateException if position is undefined at invocation time
      */
-    public LockResult previousGe(byte[] limitKey) throws IOException;
+    public default LockResult previousGe(byte[] limitKey) throws IOException {
+        return ViewUtils.previousCmp(this, limitKey, -1);
+    }
 
     /**
      * Moves to the Cursor to the previous available entry, but only when
@@ -272,7 +278,9 @@ public interface Cursor {
      * @throws NullPointerException if limit key is null
      * @throws IllegalStateException if position is undefined at invocation time
      */
-    public LockResult previousGt(byte[] limitKey) throws IOException;
+    public default LockResult previousGt(byte[] limitKey) throws IOException {
+        return ViewUtils.previousCmp(this, limitKey, 0);
+    }
 
     /**
      * Moves the Cursor to find the given key.
@@ -403,6 +411,24 @@ public interface Cursor {
      * LockResult#OWNED_EXCLUSIVE OWNED_EXCLUSIVE}
      */
     public LockResult random(byte[] lowKey, byte[] highKey) throws IOException;
+
+    /**
+     * Locks the current entry, as if by calling load. Locking is performed automatically
+     * within transactions, and so invocation of this method is necessary only when manually
+     * tweaking the lock mode.
+     *
+     * <p>By default, this method simply calls load. Subclasses are encouraged to provide a
+     * more efficient implementation.
+     *
+     * @throws IllegalStateException if position is undefined at invocation time
+     * @return {@link LockResult#UNOWNED UNOWNED}, {@link LockResult#ACQUIRED
+     * ACQUIRED}, {@link LockResult#OWNED_SHARED OWNED_SHARED}, {@link
+     * LockResult#OWNED_UPGRADABLE OWNED_UPGRADABLE}, or {@link
+     * LockResult#OWNED_EXCLUSIVE OWNED_EXCLUSIVE}
+     */
+    public default LockResult lock() throws IOException {
+        return load();
+    }
 
     /**
      * Loads or reloads the value at the cursor's current position. Cursor