+++ src/main/java/org/cojen/tupl/TrimmedCursor.java
@@ -201,6 +201,11 @@ final class TrimmedCursor implements Cursor {
     }
 
     @Override
+    public LockResult lock() throws IOException {
+        return mSource.lock();
+    }
+
+    @Override
     public LockResult load() throws IOException {
         return mSource.load();
     }